refactor(decompose): split webapp/e2e-tests/electron/critical_e2e_verification_tests/electron-markdown-editors-crud-v2.spec.ts into folder
