+++ webapp/e2e-tests/electron/critical_e2e_verification_tests/electron-markdown-editors-crud-v2/types.ts
@@ -1,0 +1,20 @@
+import type { Page } from '@playwright/test';
+import type { Core as CytoscapeCore } from 'cytoscape';
+import type { EditorView } from '@codemirror/view';
+
+export interface ExtendedWindow {
+  cytoscapeInstance?: CytoscapeCore;
+  electronAPI?: {
+    main: {
+      startFileWatching: (dir: string) => Promise<{ success: boolean; directory?: string; error?: string }>;
+      stopFileWatching: () => Promise<{ success: boolean; error?: string }>;
+      getWatchStatus: () => Promise<{ isWatching: boolean; directory?: string }>;
+    };
+  };
+}
+
+export interface CodeMirrorElement extends HTMLElement {
+  cmView?: { view: EditorView };
+}
+
+export type AppWindow = Page;
